backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==3.0.10
google-cloud-speech==2.11.1
Werkzeug==3.0.4